backend/post/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
cloudinary==1.36.0
supabase==2.0.3